+++ src/parsers/problem/HydroProblemParser.ts
@@ -10,6 +10,7 @@ export class HydroProblemParser extends Parser {
     'newoj.cyezoi.com': 'CYEZOJ',
     'oj.33dai.cn': '33OJ',
     'bs.daimayuan.top': 'Daimayuan Online Judge',
+    'yundouxueyuan.com': 'Yun Dou Xue Yuan',
   };
 
   public getMatchPatterns(): string[] {