+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for Olympicode (thanks [@EntityPlantt](https://github.com/EntityPlantt))
 - Added support for HOJ (thanks [@Introl-ljl](https://github.com/Introl-ljl))
 - Added support for Daimayuan Online Judge's bs.daimayuan.top interface
+- Added support for Yun Dou Xue Yuan
 - Fixed the SSOIER parser not working on https URLs
 
 ## [2.61.0](https://github.com/jmerle/competitive-companion/releases/tag/2.61.0) (2025-05-30)