+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed the SSOIER parser not working on https URLs
 - Fixed the TLX parser
 - Fixed the Codeforces parser omitting empty lines in sample cases
+- Fixed the KEP.uz parsers
 
 ## [2.61.0](https://github.com/jmerle/competitive-companion/releases/tag/2.61.0) (2025-05-30)
 - Added a problem parser for CSGOJ (thanks [@xiaowhang](https://github.com/xiaowhang))